requirements.txt
# Основные библиотеки
Flask==2.3.3
Flask-SQLAlchemy==3.1.1
Flask-WTF==1.2.1
WTForms==3.1.0
email-validator==2.1.0
SQLAlchemy==2.0.23

# Для работы с Word документами
python-docx==1.1.0

# Утилиты
python-dotenv==1.0.0